+++ app/pages/preview_post.tsx
@@ -13,7 +13,7 @@ import { GraphqlProvider } from "@/graphql/GraphqlProvider";
 import { defaultLocale, i18n, Locale } from "@/locales/locales";
 import { LocaleProvider } from "@/locales/use-locale";
 import { ConfiguratorStateProvider } from "@/src";
-import * as federalTheme from "@/themes/federal";
+import * as federalTheme from "@/themes/theme";
 import { migrateConfiguratorState } from "@/utils/chart-config/versioning";
 import { runMiddleware } from "@/utils/run-middleware";
